+++ src/websocket.rs
@@ -1,6 +1,5 @@
 use anyhow::{anyhow, Context, Result};
 use ed25519_dalek::*;
-use failure::{Fail, Fallible};
 use holochain::conductor::api::{
     AdminRequest, AdminResponse, AppRequest, AppResponse, InstalledAppInfo, ZomeCall,
 };
@@ -14,91 +13,19 @@ use holochain_types::{
     prelude::YamlProperties,
 };
 use holochain_websocket::{connect, WebsocketConfig, WebsocketSender};
-use hpos_config_core::{public_key, Config};
-use lazy_static::*;
-use reqwest::Client;
-use serde::*;
-use std::{collections::HashMap, env, fmt, fs, fs::File, io::prelude::*, sync::Arc};
+use hpos_config_core::{Config};
+use std::{collections::HashMap, env, fs, fs::File, io::prelude::*, sync::Arc};
 use tracing::{info, instrument, trace};
 use url::Url;
+mod membrane_proof;
 
 use crate::config::Happ;
-
-// #region structs
 
 #[derive(Clone)]
 pub struct AdminWebsocket {
     tx: WebsocketSender,
     agent_key: Option<AgentPubKey>,
 }
-
-#[derive(Debug, Fail)]
-pub enum AuthError {
-    #[fail(display = "Error: Invalid config version used. please upgrade to hpos-config v2")]
-    ConfigVersionError,
-    #[fail(display = "Registration Error: {}", _0)]
-    RegistrationError(String),
-}
-
-#[allow(non_snake_case)]
-#[derive(Debug, Serialize, Deserialize)]
-struct RegistrationError {
-    error: String,
-    isDisplayedToUser: bool,
-    info: String,
-}
-
-impl fmt::Display for RegistrationError {
-    fn fmt(&self, f: &mut fmt::Formatter) -> fmt::Result {
-        write!(f, "Error: {}, More Info: {}", self.error, self.info)
-    }
-}
-
-#[derive(Debug, Serialize)]
-struct Registration {
-    registration_code: String,
-    #[serde(serialize_with = "serialize_holochain_agent_pub_key")]
-    agent_pub_key: PublicKey,
-    email: String,
-    payload: RegistrationPayload,
-}
-
-#[derive(Debug, Serialize)]
-struct RegistrationPayload {
-    role: String,
-}
-
-#[derive(Debug, Serialize, Deserialize)]
-struct RegistrationRequest {
-    mem_proof: String,
-}
-
-// #endregion
-
-// #region static
-
-lazy_static! {
-    static ref CLIENT: Client = Client::new();
-}
-
-fn serialize_holochain_agent_pub_key<S>(
-    public_key: &PublicKey,
-    serializer: S,
-) -> Result<S::Ok, S::Error>
-where
-    S: Serializer,
-{
-    serializer.serialize_str(&public_key::to_holochain_encoded_agent_key(public_key))
-}
-
-fn mem_proof_path() -> String {
-    match env::var("MEM_PROOF_PATH") {
-        Ok(path) => path,
-        _ => "/var/lib/configure-holochain/mem-proof".to_string(),
-    }
-}
-
-// #endregion
 
 impl AdminWebsocket {
     #[instrument(err)]
@@ -116,77 +43,6 @@ impl AdminWebsocket {
             agent_key: None,
         })
     }
-
-    fn get_hpos_config(&mut self) -> Fallible<Config> {
-        let config_path = env::var("HPOS_CONFIG_PATH")?;
-        let config_json = fs::read(config_path)?;
-        let config: Config = serde_json::from_slice(&config_json)?;
-        Ok(config)
-    }
-
-    // #region membrane proof
-
-    fn mem_proof_server_url(&mut self) -> String {
-        match env::var("MEM_PROOF_SERVER_URL") {
-            Ok(url) => url,
-            _ => "https://test-membrane-proof-service.holo.host".to_string(),
-        }
-    }
-
-    async fn try_registration_auth(
-        &mut self,
-        config: &Config,
-        holochain_public_key: PublicKey,
-    ) -> Fallible<()> {
-        match config {
-            Config::V2 {
-                registration_code,
-                settings,
-                ..
-            } => {
-                let email = settings.admin.email.clone();
-                let payload = Registration {
-                    registration_code: registration_code.clone(),
-                    agent_pub_key: holochain_public_key,
-                    email: email.clone(),
-                    payload: RegistrationPayload {
-                        role: "host".to_string(),
-                    },
-                };
-                let mem_proof_server_url =
-                    format!("{}/register-user/", self.mem_proof_server_url());
-                let resp = CLIENT
-                    .post(mem_proof_server_url)
-                    .json(&payload)
-                    .send()
-                    .await?;
-                match resp.error_for_status_ref() {
-                    Ok(_) => {
-                        let reg: RegistrationRequest = resp.json().await?;
-                        println!("Registration completed message ID: {:?}", reg);
-                        // save mem-proofs into a file on the hpos
-                        let mut file = File::create(mem_proof_path())?;
-                        file.write_all(reg.mem_proof.as_bytes())?;
-                    }
-                    Err(_) => {
-                        let err: RegistrationError = resp.json().await?;
-                        return Err(AuthError::RegistrationError(err.to_string()).into());
-                    }
-                }
-            }
-            Config::V1 { .. } => {
-                return Err(AuthError::ConfigVersionError.into());
-            }
-        }
-        Ok(())
-    }
-
-    async fn enable_memproof_dev_net(&mut self, agent_key: PublicKey) -> Fallible<()> {
-        let config = self.get_hpos_config()?;
-        self.try_registration_auth(&config, agent_key).await
-    }
-
-    // #endregion
 
     #[instrument(skip(self), err)]
     pub async fn get_agent_key(&mut self) -> Result<AgentPubKey> {
@@ -249,7 +105,7 @@ impl AdminWebsocket {
                 // if using devNet,
                 // enable membrane proof using generated key
                 let agent_pub_key = PublicKey::from_bytes(key.get_raw_32())?;
-                if let Err(e) = self.enable_memproof_dev_net(agent_pub_key).await {
+                if let Err(e) = membrane_proof::enable_memproof_dev_net(agent_pub_key).await {
                     info!("membrane proof error {}", e);
                 }
                 Ok(key)
